Send email confirmation when application is rejected
