+++ applications/tests/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+from django_ilmoitin.models import NotificationTemplate
 
 from berth_reservations.tests.conftest import *  # noqa
 from berth_reservations.tests.factories import CustomerProfileFactory
@@ -8,6 +9,7 @@ from users.tests.conftest import *  # noqa
 
 from ..enums import ApplicationStatus
 from ..models import BerthSwitch, BerthSwitchReason
+from ..notifications import NotificationType
 from .factories import BerthApplicationFactory, WinterStorageApplicationFactory
 
 
@@ -72,3 +74,13 @@ def handled_ws_application(winter_storage_application):
     winter_storage_application.status = ApplicationStatus.HANDLED
     winter_storage_application.save()
     return winter_storage_application
+
+
+@pytest.fixture
+def notification_template_berth_application_rejected():
+    return NotificationTemplate.objects.language("fi").create(
+        type=NotificationType.BERTH_APPLICATION_REJECTED.value,
+        subject="test berth application rejected subject, event: {{ application.first_name }}!",
+        body_html="<b>test berth application rejected body HTML!</b>",
+        body_text="test berth application rejected body text!",
+    )
